pyproject.toml
[tool.poetry]
name = "portfolio_admin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "^3.4.1"
Django = "^3.2.7"
psycopg2 = "^2.9.1"
black = "^21.9b0"
isort = "^5.9.3"
environs = "^9.3.3"
Pillow = "^8.3.2"
drf-yasg = "^1.20.0"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
django-cors-headers = "^3.8.0"
generics = "^3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 4